% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{package}
\name{abd-package}
\alias{abd}
\alias{abd-package}
\title{Data sets from The Analysis of Biological Data}
\description{
The \code{abd} package contains data sets and sample code for the book,
\emph{The Analysis of Biological Data} by Michael C. Whitlock and Dolph
Schluter (2009; Roberts and Company Publishers).
}
\details{
\tabular{ll}{ Package: \tab abd\cr
              Type: \tab Package\cr
              Version: \tab 0.2-7\cr
              Date: \tab 2015-02-01\cr
              License: \tab GPL\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\examples{
trellis.par.set(theme=col.abd())  # set color theme
show.settings()
abdData(3)                        # look for data sets in chapter 3
abdData('Finch')                  # look for data sets with 'finch' in name
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu}); Randall Pruim
(\url{rpruim@calvin.edu})
}
\references{
Whitlock, M.C. and D. Schluter. 2009. \emph{The Analysis of
Biological Data}. Roberts and Company Publishers. ISBN: 0981519407.
\url{http://www.roberts-publishers.com/biology/the-analysis-of-biological-data.html}
}
\keyword{package}

