% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Antilles}
\alias{Antilles}
\title{Antilles Bird Immigration Dates}
\format{A data frame with 37 observations of one variable. \describe{
\item{immigration.date}{approximate immigration date (in millions of
years)} }}
\source{
\emph{inferred from} Ricklefs, R.E. and E. Bermingham. 2001.
Nonequilibrium diversity dynamics of the Lesser Antillean avifauna.
\emph{Science} 294: 1522-1524.
}
\description{
Approximate dates of immigration for 37 species of birds in the Lesser
Antilles.
}
\examples{
histogram(~immigration.date, Antilles,n=15)
densityplot(~immigration.date, Antilles)
}
\references{
\url{http://www.sciencemag.org/cgi/content/abstract/sci;294/5546/1522}
}
\keyword{datasets}

