% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Aspirin}
\alias{Aspirin}
\title{Effects of Aspirin on Cancer Rates}
\format{A data frame with 39876 observations on the following \describe{
\item{treatment}{a factor with levels \code{Aspirin} and
\code{Placebo}} \item{cancer}{a factor with levels \code{no} and
\code{yes}} }}
\source{
Cook, N.R., I. Lee, J.M. Gaziano, D. Gordon, P.M. Ridker, J.E.
Manson, C.H. Hennekens, and J.E. Buring. 2005. Low-dose aspirin in the
primary prevention of cancer. \emph{Journal of the American Medical
Association} 294: 47-55.
}
\description{
Frequency of cancer in 39,876 women taking and not taking aspirin.
}
\examples{
demo(sec9.2)
}
\references{
\url{http://jama.ama-assn.org/cgi/content/abstract/294/1/47}
}
\keyword{datasets}

