% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BeeGenes}
\alias{BeeGenes}
\title{Foraging Gene Expression}
\format{A data frame with 6 observations on the following 3 variables.
\describe{ \item{type}{a factor with levels \code{forager}
\code{nurse}} \item{colony}{a numeric identifier}
\item{expression}{expression level of the \emph{for} gene} }}
\source{
Ben-Shahar, Y., A. Robichon, M.B. Sokolowski, and G.E. Robinson.
2002. Influence of gene action across different time scales on behavior.
\emph{Science} 296: 741-744.
}
\description{
Levels of expression of the foraging gene (\emph{for}; \code{Expression}) in
two worker types (\code{type}) in three bee colonies (\code{colony}). Note
that \code{colony} is not coded as a factor.
}
\examples{
str(BeeGenes)
BeeGenes
xtabs( expression ~ type + colony, BeeGenes )
}
\keyword{datasets}

