% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Beetles}
\alias{Beetles}
\title{Beetle Wings and Horns}
\format{A data frame with 19 observations on the following 2 variables.
\describe{ \item{horn.size}{a numeric vector}
\item{wing.mass}{a numeric vector} }}
\source{
Emlen, D.J. 2001. Costs and the diversification of exaggerated
animal structures. \emph{Science} 291: 1534-1536.
}
\description{
Relative size of the horns and wings in 19 female \emph{Onthophagus
sagittarius} beetles.
}
\examples{
str(Beetles)
xyplot(wing.mass ~ horn.size, Beetles)
}
\references{
\url{http://www.scienceonline.org/cgi/content/abstract/291/5508/1534}
}
\keyword{datasets}

