% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BrainExpression}
\alias{BrainExpression}
\title{Proteolipid Protein 1 Gene Expression}
\format{A data frame with 45 observations on the following 2 variables.
\describe{ \item{group}{a factor with levels: \code{bipolar},
\code{control}, and \code{schizo}} \item{PLP1.expression}{a numeric
vector} }}
\source{
\emph{inferred from} Tkachev, D., M.L. Mimmack, M.M. Ryan, M.
Wayland, T. Freeman, P.B. Jones, M. Starkey, M.J. Webster, R.H. Yolken, S.
Bahn. 2003. Oligodendrocyte dysfunction in schizophrenia and bipolar
disorder. \emph{Lancet} 362(9386): 798-805.
}
\description{
Expression levels of the proteolipid protein 1 gene (PLP1;
\code{PLP1.expression}) in 45 individuals in one of three \code{group}s.
}
\examples{
bwplot(PLP1.expression ~ group, BrainExpression)
}
\keyword{datasets}

