% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cavalry}
\alias{Cavalry}
\title{Deaths from Horse Kicks}
\format{A data frame with 5 observations on the following 2 variables.
\describe{ \item{deaths}{a numeric vector} \item{count}{a
numeric vector} }}
\source{
Bortkiewicz, L. 1898. Das Gesetz der Kleinen Zahlen (Teubner,
Leipzig), \emph{as cited in} Larson, R.J. and M.L. Marx. 1981. \emph{An
Introduction to Mathematical Statistics and its Applications}.
Prentice-Hall: Englewood Cliffs, NJ.
}
\description{
Numbers of deaths resulting from horse kicks per regiment-years for the
Prussian army.
}
\examples{
Cavalry
xyplot(count ~ deaths, Cavalry, type='h', lwd=4)
barchart(count ~ deaths, Cavalry, horizontal = FALSE,
         box.ratio = 1000, origin=0)
}
\keyword{datasets}

