% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Clearcuts}
\alias{Clearcuts}
\title{Biomass Change in Rainforests near Clearcuts}
\format{A data frame with 36 observations of one variable. \describe{
\item{biomass.change}{} }}
\source{
Laurance, W.F., S.G. Laurance, L.V. Ferreira, J.M. Rankin-de Merona,
C. Gascon, T.E. Lovejoy. 1997. Biomass collapse in Amazonian forest
fragments. \emph{Science} 278: 1117-1118.
}
\description{
Biomass change in 36 Amazonian rainforests following clearcuts ranging from
50 m to several kilometers.
}
\examples{
str(Clearcuts)
histogram(~biomass.change, Clearcuts)
}
\references{
\url{http://www.sciencemag.org/cgi/content/abstract/278/5340/1117}
}
\keyword{datasets}

