% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CocaineDopamine}
\alias{CocaineDopamine}
\title{Effects of Cocaine on Dopamine Receptors}
\format{A data frame with 34 observations on the following 2 variables.
\describe{ \item{percent.blocked}{a numeric vector}
\item{high}{a numeric vector} }}
\source{
Volkow, N.D., G.-J. Wang, R.W. Foltin, J.S. Fowler, N.N. Abumrad, S.
Vitkun, J. Logan, S.J. Gatley, N. Pappas, R. Hitzemann, and C.E. Shea. 1997.
Relationship between subjective effects of cocaine and dopamine transporter
occupancy. \emph{Nature} 386: 827-830.
}
\description{
Percent of dopamine receptors blocked (\code{percent.bocked}) and the
perceived level of high as determined by PET scans (\code{high}) in 34
humans.
}
\examples{
str(CocaineDopamine)
xyplot(high ~ percent.blocked, CocaineDopamine)
}
\references{
\url{http://www.nature.com/nature/journal/v386/n6627/abs/386827a0.html}
}
\keyword{datasets}

