% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Convictions}
\alias{Convictions}
\title{Frequency of Convictions for a Cohort of English Boys}
\format{A data frame with 15 observations on the following 2 variables.
\describe{ \item{convictions}{number of convictions}
\item{boys}{number of boys with given number of convictions} }}
\source{
Farrington, D.P. 1994. \emph{Cambridge Study in Delinquent
Development} [Great Britain], 1961-1981. 2nd ICPSR ed. Inter-university
Consortium for Political and Social Research, Ann Arbor, MI.
}
\description{
Data on frequency of convictions for a cohort of 395 boys.
}
\examples{
str(Convictions)
barchart(boys ~ as.factor(convictions), Convictions, horizontal = FALSE, origin=0)
xyplot( boys ~ convictions, Convictions, type = "h", lwd = 20)
}
\references{
\url{http://www.icpsr.umich.edu/icpsrweb/NACJD/archive.jsp}
}
\keyword{datasets}

