% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ConvictionsAndIncome}
\alias{ConvictionsAndIncome}
\title{Convictions and Income Level in a Cohort of English Boys}
\format{A data frame with 395 observations on the following 2 variables.
\describe{ \item{convicted}{a factor with levels \code{no} and
\code{yes}} \item{income}{a factor with levels \code{adequate},
\code{comfortable}, and \code{inadequate}} }}
\source{
Farrington, D.P. 1994. \emph{Cambridge Study in Delinquent
Development} [Great Britain], 1961-1981. 2nd ICPSR ed. Inter-university
Consortium for Political and Social Research, Ann Arbor, MI.
}
\description{
Data reporting the number of individual with and without convictions per
income level.
}
\examples{
str(ConvictionsAndIncome)
ConvictionsAndIncome

xtabs(~ convicted + income, data = ConvictionsAndIncome)
}
\references{
\url{http://www.icpsr.umich.edu/icpsrweb/NACJD/archive.jsp}
}
\keyword{datasets}

