% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Crickets}
\alias{Crickets}
\title{Immunity and Sperm Viability in Crickets}
\format{A data frame with 41 observations on the following 2 variables.
\describe{ \item{sperm.viability}{a numeric vector}
\item{lysozyme}{a numeric vector} }}
\source{
Simmons, L.W. and B. Roberts. 2005. Bacterial immunity traded for
sperm viability in male crickets. \emph{Science} 309: 2031.
}
\description{
Sperm viability and immune function, measured by lysozyme activity in
crickets. Each observation is a mean for a single family of males.
}
\examples{
Crickets
xyplot(lysozyme ~ sperm.viability, Crickets)
}
\keyword{datasets}

