\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
  cv(x)
}
\arguments{
  \item{x}{A numeric vector}
}
\description{
  Compute coefficient of variation (ratio of standard
  deviation to mean).
}
\examples{
cv(GlidingSnakes$undulation.rate)
}
\author{
  Randall Pruim (\email{rpruim@calvin.edu})
}
\keyword{stats}
\keyword{univar}

