% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DesertBirds}
\alias{DesertBirds}
\title{Desert Bird Census Data}
\format{A data frame with 43 observations on the following 2 variables.
\describe{ \item{species}{a character vector} \item{count}{a
numeric vector} }}
\source{
Sauer, J.R., J.E. Hines, and J. Fallon. 2003. The North American
breeding bird survey, results and analysis 1966-2002. Version 2003.1. USGS
Patuxent Wildlife Research Center, Laurel, MD.
}
\description{
Census data for desert birds.
}
\examples{
str(DesertBirds)
histogram(~ count, DesertBirds,
  xlab = "Abundance"
  )
}
\references{
\url{http://www.mbr-pwrc.usgs.gov/bbs/}
}
\keyword{datasets}

