% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Dioecy}
\alias{Dioecy}
\title{Dioecy vs. Monomorphism in Plants}
\format{A data frame with 28 observations on the following 3 variables.
\describe{ \item{dioecious}{a numeric vector}
\item{monomorphic}{a numeric vector}
\item{taxon.pair}{identifier for pair} }}
\source{
Heilbuth, J.C. 2000. Lower species richness in dioecious clades.
\emph{The American Naturalist} 156: 221-241.
}
\description{
Number of \code{dioecious} and \code{monomorphic} taxa among pairs of
closely related plants.
}
\examples{
xyplot(dioecious ~ monomorphic, Dioecy, alpha = 0.65, pch = 16)
}
\keyword{datasets}

