% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Earwigs}
\alias{Earwigs}
\title{Earwig Density and Forceps}
\format{A data frame with 7 observations on the following 2 variables.
\describe{ \item{density}{a numeric vector}
\item{proportion.forceps}{a numeric vector} }}
\source{
Tomkins, J.L.  and G.S. Brown. 2004. Population density drives the
local evolution of a threshold dimorphism. \emph{Nature} 431: 1099-1103.
}
\description{
Earwig (\emph{Forficula auricularia}) density and the proportion of trapped
earwigs with abdominal forceps (used for fighting and courtship).
}
\examples{
xyplot(proportion.forceps ~ density, data=Earwigs, type='h', lwd=6)
}
\references{
\url{http://en.wikipedia.org/wiki/Forficula_auricularia}

\url{http://www.arkive.org/common-european-earwig/forficula-auricularia/}

\url{http://eol.org/pages/473785}
}
\keyword{datasets}

