% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EndangeredSpecies}
\alias{EndangeredSpecies}
\title{Endangered and Threatened Species}
\format{A data frame with 11 observations on the following 2 variables.
\describe{ \item{taxon}{a character vector}
\item{num.species}{a numeric vector} }}
\source{
U.S. Fish and Wildlife Service. 2001. Number of U.S. listed species
per calendar year.
}
\description{
Frequency of taxon groups on the U.S. Fish and Wildlife Service list of
endangered and threatened species (2002).
}
\examples{
str(EndangeredSpecies)
EndangeredSpecies
}
\references{
\url{http://www.fws.gov/endangered/}
}
\keyword{datasets}

