% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Fireflies}
\alias{Fireflies}
\title{Spermatophore Mass in Fireflies}
\format{A data frame with 35 observations of one variable. \describe{
\item{sp.mass}{} }}
\source{
\emph{inferred from} Cratsley, C.K. and S.M. Lewis. 2003. Female
preference for male courtship flashes in \emph{Photinus ignitus} fireflies.
\emph{Behavioral Ecology} 14: 135-140.
}
\description{
Measurements of spermatophore mass (milligrams) in 35 fireflies
(\emph{Photinus ignitus}).
}
\examples{
str(Fireflies)
histogram(~sp.mass, Fireflies, n=12)
}
\references{
\url{http://beheco.oxfordjournals.org/cgi/content/abstract/14/1/135}

\url{http://en.wikipedia.org/wiki/Firefly}
}
\keyword{datasets}

