% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FireflyFlash}
\alias{FireflyFlash}
\title{Firefly Flash Duration}
\format{A data frame with 35 observations of one variable. \describe{
\item{flash}{duration of flash (milliseconds)} }}
\source{
\emph{inferred from} Cratsley, C.K. and S.M. Lewis. 2003. Female
preference for male courtship flashes in \emph{Photinus ignitus} fireflies.
\emph{Behavioral Ecology} 14: 135-140.
}
\description{
Flash duration (measured in milliseconds) of a sample of male fireflies
(\emph{Photinus ignitus}; \emph{n} = 35).
}
\examples{
str(FireflyFlash)
histogram(~flash, FireflyFlash)
}
\keyword{datasets}

