% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FlycatcherPatch}
\alias{FlycatcherPatch}
\title{Forehead Patch Size in Collared Flycatachers}
\format{A data frame with 30 observations on the following 2 variables.
\describe{ \item{patch98}{a numeric vector} \item{patch99}{a
numeric vector} }}
\source{
Griffith, S.C. and B.C. Sheldon. 2001. Phenotypic plasticity in the
expression of a sexually selected trait: neglected components of variation.
\emph{Animal Behaviour} 61: 987-993.
}
\description{
Forehead patch size in 30 male Collared Flycatachers measured in two
consecutive years.
}
\examples{
str(FlycatcherPatch)
xyplot(patch99 ~ patch98, FlycatcherPatch)
}
\keyword{datasets}

