\docType{data}
\name{FlyTestes}
\alias{FlyTestes}
\title{Testes Size in Flies}
\format{A data frame with 8 observations on the following 2 variables.
\describe{ \item{mating}{a factor with levels \code{Monogamous}
\code{Polyandrous}} \item{testes.area}{a numeric vector} }}
\source{
  Hosken, D.J. and P.I. Ward. 2001.  Experimental evidence
  for testis size evolution via sperm competition.
  \emph{Ecology Letters} 4: 10-13.
}
\description{
  Testes size (square mm; \code{Testes.area}) in 8
  populations of common yellow dung flies
  (\emph{Scathophaga stercoraria}) with different mating
  systems (\code{Mating.system}).
}
\examples{
data(FlyTestes)
str(FlyTestes)
FlyTestes
}
\references{
  \url{http://en.wikipedia.org/wiki/Scathophaga_stercoraria}
}
\keyword{datasets}

