% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GeneRegulation}
\alias{GeneRegulation}
\title{Gene Regulation in Saccharomyces}
\format{A data frame with 26 observations on the following 2 variables.
\describe{ \item{genes.regulated}{a numeric vector}
\item{count}{a numeric vector} }}
\source{
Guelzim, N., S. Bottani, P. Bourgine and F. Képès. 2002. Topological
and causal structure of the yeast transcriptional regulatory network.
\emph{Nature Genetics} 31: 60-63.
}
\description{
Number of genes regulated by 109 regulatory genes of \emph{Saccharomyces
cerevisiae}.
}
\examples{
str(GeneRegulation)
xyplot(count ~ genes.regulated, GeneRegulation, type='h', lwd=3)
}
\keyword{datasets}

