% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Grassland}
\alias{Grassland}
\title{Grassland Diversity}
\format{A data frame with 10 observations on the following 2 variables.
\describe{ \item{nutrients}{a numeric vector}
\item{num.species}{a numeric vector} }}
\source{
Harpole, W. S. and D. Tilman. 2007. Grassland species loss due to
reduced niche dimension. \emph{Nature} 446: 791-793.
}
\description{
Species diversity in 10 experimental plots in the Park Grass Experiment at
Rothamsted Experimental Station to which varying numbers of nutrients have
been added.
}
\examples{
xyplot(num.species ~ jitter(nutrients, amount=0.1), Grassland, pch=16)
}
\references{
\url{http://www.rothamsted.ac.uk/}
}
\keyword{datasets}

