% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HornedLizards}
\alias{HornedLizards}
\title{Horn Length and Predation Status of Horned Lizards}
\format{A data frame with 184 observations on the following 2 variables.
\describe{ \item{horn.length}{a numeric vector}
\item{group}{a numeric vector} }}
\source{
Young, K.V., E.D. Brodie, Jr., and E.D. Brodie, III. 2004. How the
horned lizard got its horns. \emph{Science} 304: 65.
}
\description{
Squamosal horn length (mm; \code{horn.length}) and predation status
(\code{group}; living or killed) for 184 horned lizards (\emph{Phrynosoma
mcalli}).
}
\examples{
str(HornedLizards)

histogram(~horn.length | group, HornedLizards,
  layout=c(1,2),
  xlab="Horn Length (mm)")
}
\references{
\url{http://www.sciencemag.org/cgi/pdf_extract/304/5667/65}
}
\keyword{datasets}

