% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HumanBodyTemp}
\alias{HumanBodyTemp}
\title{Human Body Temperature}
\format{A data frame with 25 observations of one variable. \describe{
\item{temp}{body temperature (degrees F)} }}
\source{
Shoemaker, A. L. 1996. What's normal? -- Temperature, gender, and
heart rate. \emph{Journal of Statistics Education} 4(2).
}
\description{
Body temperature for 25 randomly chosen health people
}
\examples{
histogram(~temp, HumanBodyTemp)
stem(HumanBodyTemp$temp, scale = 2)
favstats(HumanBodyTemp$temp)
}
\references{
\url{http://www.amstat.org/publications/jse/v4n2/datasets.shoemaker.html}

Mackowiak, P.A., Wasserman, S.S., and Levine, M.M. 1992. A critical
appraisal of 98.6 degrees F, the upper limit of the normal body temperature,
and other legacies of Carl Reinhold August Wunderlich. \emph{Journal of the
American Medical Association} 268: 1578-1580.
}
\keyword{datasets}

