% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Hurricanes}
\alias{Hurricanes}
\title{Intense Hurricanes}
\format{A data frame with 4 observations on the following 2 variables.
\describe{ \item{hurricanes}{a numeric vector}
\item{count}{a numeric vector} }}
\source{
Blake, E.S., E.N. Rappaport, J.D. Jarrell, and C.W. Landsea. 2005.
The deadliest, costliest, and most intense United States tropical cyclones
from 1851 to 2006 (and other frequently requested hurricane facts).
\emph{NOAA Technical Memorandum NWS TPC-4}.
}
\description{
Number of hurricanes greater than or equal to category 3 on the
Safir-Simpson scale during the 20th century.
}
\examples{
Hurricanes
}
\references{
\url{http://www.aoml.noaa.gov/hrd/Landsea/Blakeetal_noaamemoApr2007.pdf}
}
\keyword{datasets}

