% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Iguanas}
\alias{Iguanas}
\title{Iguana Body Length Changes}
\format{A data frame with 64 observations of one variable. \describe{
\item{change.in.length}{} }}
\source{
Wikelski, M. and C. Thom. 2000. Marine iguanas shrink to survive El
Niño. \emph{Nature} 403: 37-38.
}
\description{
Body size change in 64 Galápagos marine iguanas (\emph{Amblyrhynchus
cristatus}) that survived the 1992-1993 El Niño event.
}
\examples{
str(Iguanas)
histogram(~ change.in.length, Iguanas, n = 10)
}
\references{
\url{http://en.wikipedia.org/wiki/Marine_iguana}
}
\keyword{datasets}

