% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{JetLagKnees}
\alias{JetLagKnees}
\title{Circadian Rhythm Phase Shift}
\format{A data frame with 22 observations on the following 2 variables.
\describe{ \item{treatment}{a factor with levels \code{control},
\code{eyes}, and \code{knee}} \item{shift}{a numeric vector} }}
\source{
\emph{inferred from} Wright, K.P., Jr. and C.A. Czeisler 2002.
Absence of circadian phase resetting in response to bright light behind the
knees. \emph{Science} 297: 571.
}
\description{
Shift in circadian rhythm (hours; \code{shift}) in three light treatments
(\code{treatment}).
}
\examples{
demo(sec15.1)
}
\references{
\url{http://www.sciencemag.org/cgi/content/full/297/5581/571}
}
\keyword{datasets}

