% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LionCubs}
\alias{LionCubs}
\title{Time to Reproduction in Female Lions}
\format{A data frame with 14 observations on the following 2 variables.
\describe{ \item{cause.of.death}{a factor with \code{accident} and
\code{infanticide}} \item{days.to.next.cub}{a numeric vector} }}
\source{
Packer, C. and A.E. Pusey. 1983. Adaptations of female lions to
infanticide by incoming males. \emph{The American Naturalist} 121: 716-728.
}
\description{
Time to reproduction (\code{Days}) based on whether death of previous cubs
was due to infanticide (\code{New}) or accidental (\code{Same}).
}
\examples{
xyplot(days.to.next.cub ~ cause.of.death, LionCubs)
}
\keyword{datasets}

