% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MassExtinctions}
\alias{MassExtinctions}
\title{Mass Extinction Frequency}
\format{A data frame with 21 observations on the following 2 variables.
\describe{ \item{num.extinctions}{a numeric vector}
\item{count}{a numeric vector} }}
\source{
Raup, D.M. and J.J. Sepkoski, Jr. 1982. Mass extinctions in the
marine fossil record. \emph{Science} 215: 1501-1503.
}
\description{
The frequency of mass extinctions in the fossil record.
}
\examples{
MassExtinctions
}
\references{
\url{http://www.sciencemag.org/cgi/content/abstract/sci;215/4539/1501}
}
\keyword{datasets}

