% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MoleRats}
\alias{MoleRats}
\title{Energy Expenditure in Mole Rats}
\format{A data frame with 35 observations on the following 3 variables.
\describe{ \item{caste}{a factor with levels \code{lazy} and
\code{worker}} \item{ln.mass}{a numeric vector}
\item{ln.energy}{a numeric vector} }}
\source{
\emph{inferred from} Scantlebury, M., J.R. Speakman, M.K.
Oosthuizen, T.J. Roper and N.C. Bennett. 2006. Energetics reveals
physiologically distinct castes in a eusocial mammal. \emph{Nature} 440:
795-797.
}
\description{
Energy expenditure (\code{ln.energy}) in two castes (\code{caste}) of
Damaraland mole rats (\emph{Cryptomys damarensis}) with body mass
(\code{ln.mass}) as a covariate.
}
\examples{
MoleRats
}
\references{
\url{http://www.nature.com/nature/journal/v440/n7085/abs/nature04578.html}
}
\keyword{datasets}

