% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NematodeLifespan}
\alias{NematodeLifespan}
\title{Effects of Trimethadione on Lifespan in Nematodes}
\format{A data frame with 200 observations on the following 2 variables.
\describe{ \item{treatment}{a factor with levels: \code{adult},
\code{larva}, \code{larva+adult}, and \code{water}}
\item{lifespan}{a numeric vector of lifespan} }}
\source{
\emph{inferred from} Evason, K., C. Huang, I. Yamben, D.F. Covey,
and K. Kornfeld. 2005. Anticonvulsant medications extend worm life-span.
\emph{Science} 307: 258-262.
}
\description{
\code{lifespan} of the nematode \emph{Caenorhabditis elegans} in control and
three experimental \code{treatment}s of the anticonvulsant drug
trimethadione.
}
\examples{
str(NematodeLifespan)
}
\references{
\url{http://www.sciencemag.org/cgi/content/abstract/307/5707/258}
}
\keyword{datasets}

