% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NeotropicalTrees}
\alias{NeotropicalTrees}
\title{Photosynthesis in Neotropical Trees}
\format{A data frame with 9 observations on the following 2 variables.
\describe{ \item{previous.fruits}{a numeric vector}
\item{photosynthetic.capacity}{a numeric vector} }}
\source{
\emph{inferred from} Wheelwright, N.T. and B.A. Logan. 2004.
Previous-year reproduction reduces photosynthetic capacity and slows
lifetime growth in females of a neotropical tree. \emph{Proceedings of the
National Academy of Sciences (USA)} 101: 8051-8055.
}
\description{
Photosynthetic capacity (\code{photosynthetic.capacity}) and number of
fruits produced in the previous season (\code{previous.fruits}) of 9 females
of \emph{Ocotea tenera}.
}
\examples{
str(NeotropicalTrees)
NeotropicalTrees
}
\references{
\url{http://www.pnas.org/content/101/21/8051.long}
}
\keyword{datasets}

