% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NorthSeaCod}
\alias{NorthSeaCod}
\title{Atlantic Cod Recruits}
\format{A data frame with 39 observations of one variable. \describe{
\item{log10.recruits}{} }}
\source{
\emph{inferred from} Beaugrand, G., K.M. Brander, J.A. Lindley, S.
Souissi, and P.C. Reid. 2003. Plankton effect on cod recruitment in the
North Sea. \emph{Nature} 426: 661-664.
}
\description{
Number (\eqn{\log_{10}}{log10} transformed) of Atlantic cod (\emph{Gadus
morhua}) that recruited (grew to catchable size) in the North Sea over a 39
years span.
}
\examples{
favstats(NorthSeaCod$log10.recruits)
}
\references{
\url{http://www.nature.com/nature/journal/v426/n6967/abs/nature02164.html}
}
\keyword{datasets}

