\docType{data}
\name{NoSmokingDay}
\alias{NoSmokingDay}
\title{No Smoking Day}
\format{A data frame with 10 observations on the following 3 variables.
\describe{ \item{year}{a numeric vector}
\item{injuries.before.NSD}{a numeric vector}
\item{injuries.on.NSD}{a numeric vector} }}
\source{
  Waters, A.J., M.J. Jarvis, and S.R. Sutton. 1998.
  Nicotine withdrawal and accident rates. \emph{Nature}
  394: 137.
}
\description{
  Number of workplace injuries on No Smoking Day
  (\code{Injuries.on.NSD}) compared to the same Wednesday
  in the previous year (\code{Injuries.before.NSD}) for
  1987-1996.
}
\examples{
data(NoSmokingDay)
NoSmokingDay
}
\references{
  \url{http://www.nosmokingday.org.uk/}
}
\keyword{datasets}

