% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PlantPersistence}
\alias{PlantPersistence}
\title{Population Persistence Times}
\format{A data frame with 16 observations on the following 2 variables.
\describe{ \item{generations}{a numeric vector}
\item{treatment}{a factor with levels: \code{Isolated},
\code{Medium}, \code{Long}, and \code{Continuous}} }}
\source{
Molofsky, J. and J.-B. Ferdy. 2005. Extinction dynamics in
experimental metapopulations. \emph{Proceedings of the National Academy of
Sciences (USA)} 102: 3726-3731.
}
\description{
Persistence times (\code{generations}) in the annual plant \emph{Cardamine
pensylvanica} in four experimental populations (\code{treatment}).
}
\examples{
xyplot(generations~treatment, PlantPersistence)
}
\keyword{datasets}

