% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SagebrushCrickets}
\alias{SagebrushCrickets}
\title{Sagebrush Cricket Mating Times}
\format{A data frame with 24 observations on the following 2 variables.
\describe{ \item{treatment}{a factor with levels: \code{fed} and
\code{starved}} \item{time.to.mating}{a numeric vector} }}
\source{
Chadwick Johnson, J., T.M. Ivy, and S.K. Sakaluk. 1999. Female
remating propensity contingent on sexual cannibalism in sagebrush crickets,
\emph{Cyphoderris strepitans}: a mechanism of cryptic female choice.
\emph{Behavioral Ecology} 10: 227-233.
}
\description{
Time to mating (\code{time.to.mating}) in fed and unfed (\code{treatment})
sagebrush crickets (\emph{Cyphoderris strepitans}).
}
\examples{
SagebrushCrickets
str(SagebrushCrickets)
}
\keyword{datasets}

