\docType{data}
\name{Seedlings}
\alias{Seedlings}
\title{Number of Seedlings Per Quadrat}
\format{A data frame with 8 observations on the following 2 variables.
\describe{ \item{seedlings}{a numeric vector} \item{count}{a
numeric vector} }}
\description{
  Data on frequency of seeding per quadrat for 80
  hypothetical quadrats.
}
\examples{
data(Seedlings)
Seedlings
}
\keyword{datasets}

