\docType{data}
\name{ShadParasites}
\alias{ShadParasites}
\title{Shad Parasites}
\format{A data frame with 7 observations on the following 2 variables.
\describe{ \item{parasites}{a numeric vector} \item{count}{a
numeric vector} }}
\source{
  Shaw, D.J., B.T. Grenfell, and A.P. Dobson. 1998.
  Patterns of macroparasite aggregation in wildlife host
  populations. \emph{Parasitology} 117: 597-610.
}
\description{
  Frequency of the nematode \emph{Camallanus oxycephalus}
  per fish.
}
\examples{
data(ShadParasites)
ShadParasites
}
\references{
  \url{http://journals.cambridge.org/action/displayAbstract?aid=24297}
}
\keyword{datasets}

