% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Silversword}
\alias{Silversword}
\title{Rate of Speciation in Silverswords}
\format{A data frame with 21 observations on the following variable.
\describe{ \item{waiting.time}{a numeric vector} }}
\source{
\emph{inferred from} Baldwin, B. G. and M. J. Sanderson 1998. Age
and rate of diversification of the Hawaiian silversword alliance
(Compositae). \emph{Proceedings of the National Academy of Sciences (USA)}
95: 9402-9406.
}
\description{
Speciation "waiting times" in Hawaiian silverswords (\emph{Dubautia}).
}
\examples{
Silversword
}
\keyword{datasets}

