% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SleepAndPerformance}
\alias{SleepAndPerformance}
\title{Sleep and Learning}
\format{A data frame with 10 observations on the following 2 variables.
\describe{ \item{sleep}{a numeric vector}
\item{improvement}{a numeric vector} }}
\source{
Huber, R., M.F. Ghilardi, M. Massimini, and G. Tononi. 2004. Local
sleep and learning. \emph{Nature} 430: 78-81.
}
\description{
The increase in "slow-wave" sleep and improvements in spatial learning tasks
in 10 humans.
}
\examples{
str(SleepAndPerformance)
xyplot(improvement ~ sleep, SleepAndPerformance)
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/15184907}

\url{http://www.nature.com/nature/journal/v430/n6995/full/nature02663.html}
}
\keyword{datasets}

