% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SockeyeFemales}
\alias{SockeyeFemales}
\title{Body Masses of Female Sockeye Salmon}
\format{A data frame with 228 observations of a single variable. \describe{
\item{mass}{body mass (kg)} }}
\source{
Hendry, A.P., O.K. Berg, and T.P. Quinn. 1999. Condition dependence
and adaptation-by-time: Breeding date, life history, and energy allocation
within a population of salmon. \emph{Oikos} 85: 499-514.
}
\description{
Body Masses of 228 female Sockeye Salmon (\emph{Oncorhynchus nerka};
\url{http://www.nmfs.noaa.gov/pr/species/fish/sockeyesalmon.htm})
}
\examples{
str(SockeyeFemales)
summary(SockeyeFemales)
}
\references{
\url{http://www.jstor.org/stable/3546699}
}
\keyword{datasets}

