% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Sparrows}
\alias{Sparrows}
\title{Lifetime Reproductive Success in House Sparrows}
\format{A data frame with 9 observations on the following 3 variables.
\describe{ \item{lifetimeRS}{a numeric vector}
\item{females}{a numeric vector} \item{males}{a numeric vector} }}
\source{
Jensen, H., B.-E. Saether, T.H. Ringsby, J. Tufto, S.C.
Griffith, and H. Ellegren. 2004. Lifetime reproductive success in relation
to morphology in the House Sparrow \emph{Passer domesticus}. \emph{Journal
of Animal Ecology} 73: 599-611.
}
\description{
A cross table of lifetime reproductive success (\code{LifetimeRS}) in female
and male house sparrows \emph{Passer domesticus} in Norway.
}
\examples{
Sparrows
}
\keyword{datasets}

