% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SpiderColonies}
\alias{SpiderColonies}
\title{Social Spiders}
\format{A data frame with 17 observations on the following 3 variables.
\describe{ \item{colony}{identifier} \item{height}{height of
web above ground (cm)} \item{spiders}{number of spiders in colony} }}
\source{
Rypstra, A. L. 1979. Foraging folks of spiders, a study of aggregate
behavior in \emph{Cryptophora citricola} Forskal (Araneae: Araneidae) in
West Africa. \emph{Behavioral Ecology and Sociobiology} 5: 291-300.
}
\description{
Web height above ground (cm) and colony size for 17 colonies of the spider
\emph{Cryptophora citricola} in Gabon.
}
\examples{
str(SpiderColonies)
SpiderColonies
}
\keyword{datasets}

