% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Stalkies2}
\alias{Stalkies2}
\title{Stalk-eyed Fly Eyespan}
\format{A data frame with 45 observations on the following 2 variables.
\describe{ \item{food}{a factor with levels \code{Corn}
\code{Cotton}} \item{eye.span}{a numeric vector} }}
\source{
David, P., T. Bjorksten, K. Fowler, and A. Pomiankowski. 2000.
Condition-dependent signalling of genetic variation in stalk-eyed flies.
\emph{Nature} 406: 186-188.
}
\description{
Eyespan width (mm; \code{Eye.span}) in 45 stalk-eyed flies
(\emph{Cyrtodiopsis dalmanni}) fed a corn or cotton diet (\code{Food}).
}
\examples{
str(Stalkies2)
xyplot(eye.span ~ food, Stalkies2)
aggregate(eye.span ~ food, Stalkies2, FUN = favstats)
}
\references{
\url{http://www.nature.com/nature/journal/v406/n6792/full/406186a0.html}
}
\keyword{datasets}

