% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SyrupSwimming}
\alias{SyrupSwimming}
\title{Syrup Swimming}
\format{A data frame with 18 observations of one variable. \describe{
\item{relative.speed}{ratio of speed in syrup to speed in water} }}
\source{
Gettelfinger, B. and E. L. Cussler. 2004. Will Humans Swim Faster or
Slower in Syrup? \emph{AIChE Journal} 50: 2646-2647.
}
\description{
Relative swimming speed (speed in syrup / speed in water) for 18 swimmers.
}
\examples{
SyrupSwimming
histogram(~ relative.speed, SyrupSwimming)
dotplot(~ relative.speed, SyrupSwimming)
}
\references{
\url{http://www3.interscience.wiley.com/journal/109665380/issue}
}
\keyword{datasets}

