% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Telomeres}
\alias{Telomeres}
\title{Telomere Shortening}
\format{A data frame with 39 observations on the following 2 variables.
\describe{ \item{years}{a numeric vector}
\item{telomere.length}{a numeric vector} }}
\source{
Epel, E.S., E.H. Blackburn, J. Lin, F.S. Dhabhar, N.E. Adler, J.D.
Morrow, and R.M. Cawthon. 2004. Accelerated telomere shortening in response
to life stress. \emph{Proceedings of the National Academy of Sciences (USA)}
101: 17312-17315.
}
\description{
Telomere length (ratio) and years since their child's diagnosis with chronic
illness.
}
\examples{
xyplot(years ~ telomere.length, Telomeres,
  xlab = "Time since diagnosis (years)",
  ylab = "Telomere length (ratio)"
)
}
\references{
\url{http://www.pnas.org/content/101/49/17312}
}
\keyword{datasets}

