% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/themes.R
\name{col.abd}
\alias{col.abd}
\alias{col.abd,theme.abd}
\alias{theme.abd}
\title{Lattice theme for Analysis of Biological Data}
\usage{
col.abd(bw = FALSE, lty = 1:7)

theme.abd(bw = FALSE, lty = 1:7)
}
\arguments{
\item{bw}{a logical. Use a grayscale theme instead of color?}

\item{lty}{line types used for \code{panel.superpose}}
}
\value{
a list that can be used as a \code{lattice} theme.
}
\description{
This theme will help produce plots with color scheme similar to the one used
in \emph{Analysis of Biological Data}
}
\details{
\code{theme.abd} and \code{col.abd} are the same function with
two names.
}
\examples{
trellis.par.set(theme=col.abd(bw=TRUE))
show.settings()
trellis.par.set(theme=theme.abd(lty=1))
show.settings()
}
\author{
Randall Pruim (\email{rpruim@calvin.edu})
}
\keyword{graphics}

