% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TimeOfDeath}
\alias{TimeOfDeath}
\title{Hypoxanthine and Time Since Death}
\format{A data frame with 48 observations on the following 2 variables.
\describe{ \item{hours}{a numeric vector}
\item{hypoxanthine}{a numeric vector} }}
\source{
James, R.A., P.A. Hoadley, and B.G. Sampson. 1997. Determination of
postmortem interval by sampling vitreous humor. \emph{American Journal of
Forensic Medicine and Pathology} 18: 158-162.
}
\description{
Hypoxanthine levels in the vitreous humour of the eye and time since death
(hours) for 48 subjects.
}
\examples{
xyplot(hypoxanthine ~ hours, TimeOfDeath, type=c('p','r'))
}
\keyword{datasets}

