% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Tobacco}
\alias{Tobacco}
\title{Flower Length in Tobacco Plants}
\format{A data frame with 13 observations on the following 3 variables.
\describe{ \item{flower.length}{a numeric vector of flower length in
mm} \item{f1.count}{a numeric vector of the number of F1 plants with
flower lengths in this size range} \item{f2.count}{a numeric vector
of the number of F2 plants with flower lengths in this size range} }}
\source{
East, E.M. 1916. Studies on size inheritance in \emph{Nicotiana}.
\emph{Genetics} 1: 164-176.
}
\description{
Distribution of flow lengths in F1 and F2 populations of \emph{Nicotiana}.
}
\examples{
Tobacco
}
\references{
\url{http://www.genetics.org/content/vol1/issue2/}

\url{http://en.wikipedia.org/wiki/Nicotiana}
}
\seealso{
\code{\link{Tobacco2}}
}
\keyword{datasets}

