% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ToothAge}
\alias{ToothAge}
\title{Radioactive Teeth}
\format{A data frame with 20 observations on the following 2 variables.
\describe{ \item{actual}{a numeric vector}
\item{estimated}{a numeric vector} }}
\source{
Spalding, K.L., B.A. Buchholz, L.-E. Bergman, H. Druid, and J.
Frisén. 2005. Age written in teeth by nuclear tests. \emph{Nature} 437:
333-334.
}
\description{
Actual birth year and birth year estimated from relative radioactivity of
the enamel for 20 samples.
}
\examples{
str(ToothAge)
xyplot(actual ~ estimated, ToothAge)
}
\references{
\url{http://www.nature.com/nature/journal/v437/n7057/full/437333a.html}
}
\keyword{datasets}

