% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Trematodes}
\alias{Trematodes}
\title{Frequencies of Fish Eaten by Trematode Infection Level}
\format{A data frame with 141 observations on the following 2 variables.
\describe{ \item{infection.status}{a factor with levels:
\code{high}, \code{light}, and \code{uninfected}} \item{eaten}{a
factor with levels: \code{no} and \code{yes}} }}
\source{
Lafferty, K.D. and A.K. Morris. 1996. Altered behavior of
parasitized killifish increases susceptibility to predation by bird final
hosts. \emph{Ecology} 77: 1390-1397.
}
\description{
Frequencies of killifish (\emph{Fundulus parvipinnis}) eaten by birds
depending on level of infection by the trematode \emph{Euhaplorchis
californiensis}.
}
\examples{
demo(sec9.3)
}
\keyword{datasets}

