% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Trillium}
\alias{Trillium}
\title{Trillium Recruitment near Clearcuts}
\format{A data frame with 8 observations on the following 3 variables.
\describe{ \item{population}{a numeric vector}
\item{edge.dist}{a numeric vector} \item{recruitment}{a
numeric vector} }}
\source{
Jules, E.S. and B.J. Rathcke. 1999.  Mechanisms of reduced trillium
recruitment along edges of old-growth forest fragments. \emph{Conservation
Biology} 13:784-793
}
\description{
Recruitment of \emph{Trillium} and distance to nearest clearcut in eight
populations in southwestern Oregon.
}
\examples{
str(Trillium)
splom(Trillium)
}
\references{
\url{http://www.jstor.org/stable/2641693}
}
\keyword{datasets}

