% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Vines}
\alias{Vines}
\title{Climbing Vines}
\format{A data frame with 48 observations on the following 2 variables.
\describe{ \item{climbing}{a numeric vector}
\item{nonclimbing}{a numeric vector} }}
\source{
Gianoli, E. 2004. Evolution of a climbing habit promotes
diversification in flowering plants. \emph{Proceedings of the Royal Society
of London, Series B, Biological Sciences} 271: 2011-2015.
}
\description{
Number of \code{climbing} and \code{nonclimbing} species within closely
related general of plants.
}
\examples{
xyplot(nonclimbing ~ climbing, Vines, scales=list(log=TRUE))
}
\references{
\url{http://rspb.royalsocietypublishing.org/content/271/1552/2011.full.pdf}
}
\keyword{datasets}

