% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{VoleDispersal}
\alias{VoleDispersal}
\title{Home Range Size in Field Voles}
\format{A data frame with 5 observations on the following 3 variables.
\describe{ \item{homeranges}{a numeric vector}
\item{count}{a numeric vector} \item{sex}{a factor with
levels \code{female} and \code{male}} }}
\source{
Sandell, M., J. Agrell, S. Erlinge, and J. Nelson. 1991. Adult
philopatry and dispersal in the field vole \emph{Microtus agrestis}.
\emph{Oecologia} 86: 153-158.
}
\description{
Home range size size in field voles (\emph{Microtus agrestis}).
}
\examples{
xtabs(count~sex+homeranges,VoleDispersal)
barchart( xtabs(count~sex+homeranges,VoleDispersal), origin=0, auto.key=TRUE)
barchart(count~sex+homeranges,VoleDispersal, origin=0)
barchart(count~sex,groups=homeranges,VoleDispersal, origin=0)
barchart(count~sex,groups=homeranges,VoleDispersal, origin=0,stack=TRUE)
}
\keyword{datasets}

