% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WolfTeeth}
\alias{WolfTeeth}
\title{Wolf Tooth Measurements}
\format{A data frame with 35 observations of one variable.  \describe{
\item{length}{distance from canine to last molar teach (cm)} }}
\source{
Whitlock, M. 1996. The heritability of fluctuating asymmetry and the
genetic control of developmental stability. \emph{Proceedings of the Royal
Society, Series B} 263: 849-853.
}
\description{
Measurement (cm) of the distance between the canine and last molar teeth in
35 wolves.
}
\examples{
histogram(~ length, WolfTeeth)
}
\references{
\url{http://rspb.royalsocietypublishing.org/content/263/1372/849.abstract}
}
\keyword{datasets}

