% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{YeastGenes}
\alias{YeastGenes}
\title{Yeast Regulatory Genes}
\format{A data frame with 6 observations on the following 2 variables.
\describe{ \item{genes.controlled}{a numeric vector}
\item{count}{a numeric vector} }}
\source{
Guelzim, N., S. Bottani, P. Bourgine and F. Képès. 2002. Topological
and causal structure of the yeast transcriptional regulatory network.
\emph{Nature Genetics} 31: 60-63.
}
\description{
Number of genes regulated by 109 yeast regulatory genes.
}
\examples{
str(YeastGenes)
barchart(count ~ genes.controlled , origin=0, YeastGenes, horizontal=FALSE)
}
\keyword{datasets}

