% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ZebraFinchBeaks}
\alias{ZebraFinchBeaks}
\title{Mate Preference in Zebra Finches}
\format{A numeric vector with 10 observations.}
\source{
Blount, J.D., N.B. Metcalfe, T.R. Birkhead, P.F. Surai. 2003.
Carotenoid modulation of immune function and sexual attractiveness in Zebra
Finches. \emph{Science} 300: 125-127.
}
\description{
Percentage of time that a female spent next to a carotenoid-supplemented
male Zebra Finch compared to his non-supplemented brother.
}
\examples{
ZebraFinchBeaks
}
\references{
\url{http://www.sciencemag.org/cgi/content/abstract/300/5616/125}
}
\keyword{datasets}

