% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ZebraFinches}
\alias{ZebraFinches}
\title{Zebra Finch Carotenoids}
\format{A data frame with 20 observations on the following 3 variables.
\describe{ \item{treatment}{a factor with levels: \code{CAROT} and
\code{NO}} \item{PHA}{a numeric vector} \item{SRBC}{a
numeric vector} }}
\source{
McGraw, K.J. and D.R. Ardia. 2003. Carotenoids, immunocompetence,
and the information content of sexual colors: an experimental test.
\emph{The American Naturalist} 162: 704-712.
}
\description{
Data on cell-mediated immunocompetence (\code{PHA}) and humoral immunity
(\code{SRBC}) in Zebra Finches that received supplemental carotenoids
(\code{CAROT}) and those that did not (\code{NO}).
}
\examples{
ZebraFinches
}
\keyword{datasets}

