% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ZooMortality}
\alias{ZooMortality}
\title{Home Range Size and Mortality}
\format{A data frame with 20 observations on the following 2 variables.
\describe{ \item{log.homerange}{a numeric vector}
\item{mortality}{a numeric vector} }}
\source{
Clubb, R. and G. Mason. 2003. Captivity effects on wide ranging
carnivores. \emph{Nature} 425: 473-474.
}
\description{
Home range size (\eqn{\log_{10}}{log10}) and captive infant mortality (%)
for 20 species of carnivores.
}
\examples{
str(ZooMortality)
}
\references{
\url{http://www.nature.com/nature/journal/v425/n6957/full/425473a.html}
}
\keyword{datasets}

