\name{abd-package}
\alias{abd-package}
\alias{abd}
\docType{package}
\title{
abd: data sets from The Analysis of Biological Data
}

\description{
The \code{abd} package contains data sets and sample code for the book, \emph{The Analysis of Biological Data} by Michael C. Whitlock and Dolph Schluter (2009; Roberts and Company Publishers).
}

\details{
	\tabular{ll}{
	Package: \tab abd\cr
	Type: \tab Package\cr
	Version: \tab 0.1\cr
	Date: \tab 2009-12-06\cr
	License: \tab GPL\cr
	LazyLoad: \tab yes\cr
}

An overview of how to use the package, including the most important functions

Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
}

\author{
Kevin M. Middleton \url{kmm@csusb.edu}

Maintainer: Kevin M. Middleton \url{kmm@csusb.edu}
}

\references{
Whitlock, M.C. and D. Schluter. 2009. \emph{The Analysis of Biological Data}. Roberts and Company Publishers. ISBN: 0981519407. \url{http://www.roberts-publishers.com/whitlock/}
}

\keyword{ package }

%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}

\examples{
example(DesertBirds)
}
