\name{abd-package}
\alias{abd-package}
\alias{abd}
\docType{package}
\title{Data sets from The Analysis of Biological Data}

\description{
The \code{abd} package contains data sets and sample code for the book, \emph{The Analysis of Biological Data} by Michael C. Whitlock and Dolph Schluter (2009; Roberts and Company Publishers).
}

\details{
  \tabular{ll}{
	Package: \tab abd\cr
	Type: \tab Package\cr
	Version: \tab 0.1-10\cr
	Date: \tab 2011-1-8\cr
	License: \tab GPL\cr
	LazyLoad: \tab yes\cr
}

\tabular{ll}{
Overview\tab \code{file.show(system.file("README", package = "abd"))}\cr
News\tab \code{RShowDoc("NEWS", package = "abd")}\cr
Acknowledgements\tab \code{RShowDoc("THANKS", package = "abd")}\cr
This File\tab \code{?abd}\cr
}

}

\author{
Kevin M. Middleton (\url{kmm@csusb.edu}); Randall Pruim (\url{rpruim@calvin.edu})
}

\references{Whitlock, M.C. and D. Schluter. 2009. \emph{The Analysis of Biological Data}. Roberts and Company Publishers. ISBN: 0981519407. \url{http://www.roberts-publishers.com/whitlock/}}

\keyword{ package }

%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}

%\examples{
%}
