\name{AntillesImmigrationDates}
\alias{AntillesImmigrationDates}
\docType{data}

\title{
Antilles Bird Immigration Dates
}

\description{
Approximate dates of immigration for 37 species of birds in the Lesser Antilles.
}

\usage{data(AntillesImmigrationDates)}
\format{
  The format is:
 num [1:37] 0 0 0 0.2 0.3 0.5 0.6 0.9 1 1.3 ...
}

% \details{
% }

\source{
\emph{inferred from} Ricklefs, R.E. and E. Bermingham. 2001. Nonequilibrium diversity dynamics of the Lesser Antillean avifauna. \emph{Science} 294: 1522-1524.
}

\references{
\url{http://www.sciencemag.org/cgi/content/abstract/sci;294/5546/1522}
}

\examples{
data(AntillesImmigrationDates)
AntillesImmigrationDates
}

\keyword{datasets}
