\name{cv}
\alias{cv}

\title{Coefficient of Variation}
\description{
Compute coefficient of variation (ratio of standard deviation to mean).
}

\usage{
cv(x)
}

\arguments{
  \item{x}{A numeric vector}
}


\author{
Randall Pruim (\email{rpruim@calvin.edu})
}

\examples{
cv(GlidingSnakes$undulation.rate)
}

\keyword{ univar }
\keyword{ stats }
