% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataInfo}
\alias{dataInfo}
\title{\code{abd} Data Sets}
\format{
A data frame with 143 observations on the following 5 variables.
\describe{ \item{name}{name of data set}
\item{chapter}{chapter in which data set appears}
\item{type}{used in an \code{Example} or a \code{Problem}}
\item{number}{example or problem number}
\item{sub}{sub-problem: \code{} \code{a} \code{b} \code{c}} }
}
\description{
Information about the location of data sets in \emph{Analysis of Biological
Data}
}
\examples{

str(dataInfo)

}
\seealso{
\code{\link{abdData}}
}
\keyword{datasets}
