% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Newts}
\alias{Newts}
\title{Tetrodotoxin Resistance in Garter Snakes}
\format{A data frame with 12 observations on the following 2 variables.
\describe{ \item{locality}{a factor with levels: \code{Benton} and
\code{Warrenton}} \item{resistance}{a numeric vector} }}
\source{
Geffeney, S., E.D. Brodie, Jr., P.C. Ruben, and E.D. Brodie III.
2002. Mechanisms of adaptation in a predator-prey arms race: TTX-resistant
sodium channels. \emph{Science} 297: 1336-1339.
}
\description{
Percent reduction in crawl speed (\code{resistance}) in the garter snake
after injection of the neurotoxin tetrodotoxin from the rough-skinned newt
(\emph{Taricha granulosa}).
}
\examples{
Newts
}
\references{
\url{http://www.sciencemag.org/cgi/content/abstract/297/5585/1336}
}
\keyword{datasets}

