% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Seedlings}
\alias{Seedlings}
\title{Number of Seedlings Per Quadrat}
\format{A data frame with 8 observations on the following 2 variables.
\describe{ \item{seedlings}{a numeric vector} \item{count}{a
numeric vector} }}
\description{
Data on frequency of seeding per quadrat for 80 hypothetical quadrats.
}
\examples{
Seedlings
}
\keyword{datasets}

