% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ShadParasites}
\alias{ShadParasites}
\title{Shad Parasites}
\format{A data frame with 7 observations on the following 2 variables.
\describe{ \item{parasites}{a numeric vector} \item{count}{a
numeric vector} }}
\source{
Shaw, D.J., B.T. Grenfell, and A.P. Dobson. 1998. Patterns of
macroparasite aggregation in wildlife host populations. \emph{Parasitology}
117: 597-610.
}
\description{
Frequency of the nematode \emph{Camallanus oxycephalus} per fish.
}
\examples{
ShadParasites
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/9881385}
}
\keyword{datasets}

