\docType{data}
\name{Stalkies1}
\alias{Stalkies1}
\title{Eye Widths in Stalk-Eyed Flies}
\format{a data frame with 9 observations of 1 variable \describe{
\item{eye.span}{eye span (mm)} }}
\source{
  Data provided by Kevin Fowler, University College,
  London.
}
\description{
  Eye width in 9 male stalk-eyed flies (\emph{Cyrtodiopsis
  dalmanni}).
}
\examples{
data(Stalkies1)
Stalkies1
}
\keyword{datasets}

