\docType{data}
\name{WalkingStickFemurs}
\alias{WalkingStickFemurs}
\title{Walking Stick Femur Length}
\format{A data frame with 50 observations on the following 2 variables.
\describe{ \item{specimen}{a integer denoting specimen number.}
\item{femur.length}{a numeric vector of femur length} }}
\source{
  Nosil, P. and B.J. Crespi. 2006. Experimental evidence
  that predation promotes divergence in adaptive radiation.
  \emph{Proceedings of the National Academy of Sciences
  (USA)} 103: 9090-9095.
}
\description{
  Two measures of femur length \code{femur.length} for each
  of 25 walking sticks (\emph{Timema cristinae}). Note that
  \code{specimen} is not coded as a factor.
}
\examples{
demo(sec15.6)
}
\references{
  \url{http://www.sfu.ca/biology/faculty/crespi/pdfs/96-Nosil&Crespi2006PNAS.pdf}
}
\keyword{datasets}

