% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WrasseSexes}
\alias{WrasseSexes}
\title{Distribution of Wrasses}
\format{A data frame with 3 observations on the following 3 variables.
\describe{ \item{males}{a numeric vector} \item{females}{a
numeric vector} \item{count}{a numeric vector} }}
\description{
Number and sex of adult wrasses in a section of the Great Barrier Reef.
}
\examples{
xtabs(count ~ males + females, WrasseSexes)
}
\keyword{datasets}

